Refactor the 'service' folder
