+++ srcs/frontend/static/js/services/i18next.js
@@ -1,5 +1,5 @@
-import { renderPage } from "/static/js/services/index.js";
 import { LOCALES } from "/static/variables.js";
+import { renderPage } from "./index.js";
 
 const i18nextInit = async () => {
   await Promise.all(
@@ -25,17 +25,17 @@ const i18nextInit = async () => {
         });
     }),
   )
-    .then((localesData) => {
-      i18next.init({
-        lng: LOCALES[0],
-        fallbackLng: LOCALES[0],
-        resources: localesData.reduce((acc, curr) => ({ ...acc, ...curr }), {}),
-      });
-      i18next.on("languageChanged", renderPage);
-    })
-    .catch((error) => {
-      throw new Error(`Failed to initiate the i18next: ${error}`);
+  .then((localesData) => {
+    i18next.init({
+      lng: LOCALES[0],
+      fallbackLng: LOCALES[0],
+      resources: localesData.reduce((acc, curr) => ({ ...acc, ...curr }), {}),
     });
+    i18next.on("languageChanged", renderPage);
+  })
+  .catch((error) => {
+    throw new Error(`Failed to initiate the i18next: ${error}`);
+  });
 };
 
 export default i18nextInit;
