+++ srcs/frontend/static/js/services/index.js
@@ -1,7 +1,8 @@
 import generateSocket from "./generate-socket.js";
 import i18nextInit from "./i18next.js";
 import navigateTo from "./navigation.js";
-import renderPage from "./render.js";
 import sendNotification from "./notifications.js";
+import renderPage from "./render.js";
+import chatMessageSocketListener from "./socket-listeners.js";
 
-export { generateSocket, i18nextInit, navigateTo, renderPage, sendNotification };
+export { generateSocket, i18nextInit, navigateTo, sendNotification, renderPage, chatMessageSocketListener };
